Moved to CustomPostTypes namespace
