+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Registries/CustomPostTypeRegistry.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Registries;
 
-use GraphQLAPI\GraphQLAPI\Services\PostTypes\CustomPostTypeInterface;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\CustomPostTypeInterface;
 
 class CustomPostTypeRegistry implements CustomPostTypeRegistryInterface
 {
