+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/BlockCategories/AccessControlBlockCategory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI\Services\BlockCategories;
 
 use GraphQLAPI\GraphQLAPI\Services\BlockCategories\AbstractBlockCategory;
-use GraphQLAPI\GraphQLAPI\Services\PostTypes\GraphQLAccessControlListPostType;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLAccessControlListPostType;
 
 class AccessControlBlockCategory extends AbstractBlockCategory
 {
