+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/BlockCategories/PersistedQueryBlockCategory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\BlockCategories;
 
-use GraphQLAPI\GraphQLAPI\Services\PostTypes\GraphQLPersistedQueryPostType;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLPersistedQueryPostType;
 
 class PersistedQueryBlockCategory extends AbstractBlockCategory
 {
