+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/BlockCategories/SchemaConfigurationBlockCategory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\BlockCategories;
 
-use GraphQLAPI\GraphQLAPI\Services\PostTypes\GraphQLSchemaConfigurationPostType;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLSchemaConfigurationPostType;
 
 class SchemaConfigurationBlockCategory extends AbstractBlockCategory
 {
