+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/Clients/CustomEndpointClientTrait.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI\Services\Clients;
 
 use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
-use GraphQLAPI\GraphQLAPI\Services\PostTypes\GraphQLEndpointPostType;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLEndpointPostType;
 use PoP\ComponentModel\Misc\RequestUtils;
 
 trait CustomEndpointClientTrait
